First try to refactor pdf format
