+++ vaadin-grid-exporter/src/main/java/software/xdev/vaadin/grid_exporter/ReportException.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright © 2022 XDEV Software (https://xdev.software/en)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package software.xdev.vaadin.grid_exporter;
+
+
+
+public class ReportException extends RuntimeException
+{
+	public ReportException()
+	{
+		super();
+	}
+	
+	public ReportException(final String message, final Throwable cause)
+	{
+		super(message, cause);
+	}
+	
+	public ReportException(final String message)
+	{
+		super(message);
+	}
+	
+	public ReportException(final Throwable cause)
+	{
+		super(cause);
+	}
+}
