+++ vaadin-grid-exporter/src/main/java/software/xdev/vaadin/grid_exporter/Translator.java
@@ -1,0 +1,7 @@
+package software.xdev.vaadin.grid_exporter;
+
+@FunctionalInterface
+public interface Translator
+{
+	String translate(final String key);
+}
