+++ vaadin-grid-exporter/src/main/java/software/xdev/vaadin/grid_exporter/format/FormatConfigComponent.java
@@ -1,0 +1,9 @@
+package software.xdev.vaadin.grid_exporter.format;
+
+import com.vaadin.flow.component.orderedlayout.FlexComponent;
+
+
+public interface FormatConfigComponent<T extends SpecificConfig> extends FlexComponent
+{
+	T getConfig();
+}
