+++ vaadin-grid-exporter/src/main/java/software/xdev/vaadin/grid_exporter/format/SpecificConfig.java
@@ -1,0 +1,5 @@
+package software.xdev.vaadin.grid_exporter.format;
+
+public interface SpecificConfig
+{
+}
